Collect class names first, add to node after
This is a big change. Previously when walking the tree, class names were added to each node as walked. This became an issue when adding descendant classes (:first-child etc) as those were actually added to the children of the current node, which lead to issue with class name order. Now class names to be added are stored temporarily in an array to , then added on a second walk. Yes, the tree is walked twice. Yes this is dumb.